refactor: move Textarea to app/components reusing Input field shell; reject feedback submit in browser preview
